needs route testing

+++ api/util/imagery.go
@@ -227,21 +227,21 @@ func ImageFromBands(paths []string, ramp []uint8, transform func(...uint16) floa
 }
 
 // ScaleConfidenceMatrix scales confidence matrix to desired size using linear scaling
-func ScaleConfidenceMatrix(width int, height int, confidence *[][]float32) [][]float32 {
-	resultMatrix := make([][]float32, height)
+func ScaleConfidenceMatrix(width int, height int, confidence *[][]float64) [][]float64 {
+	resultMatrix := make([][]float64, height)
 	confWidth := len((*confidence)[0]) - 1
 	adjustedWidth := width - 1 //adjust width to work with arrays
 	// create 2d matrix
 	for i := range resultMatrix {
-		resultMatrix[i] = make([]float32, width)
+		resultMatrix[i] = make([]float64, width)
 	}
 	// first lerp columns, then using the column data lerp the rows
 	columns := getConfidenceColumns(height, confidence)
 	for y := 0; y < len(columns); y++ {
 		for x := 0; x < confWidth; x++ {
 			next := x + 1
-			xStartPos := int(lerp(0.0, float32(adjustedWidth), float32(x)/float32(confWidth)))
-			xEndPos := int(lerp(0.0, float32(adjustedWidth), float32(next)/float32(confWidth)))
+			xStartPos := int(lerp(0.0, float64(adjustedWidth), float64(x)/float64(confWidth)))
+			xEndPos := int(lerp(0.0, float64(adjustedWidth), float64(next)/float64(confWidth)))
 			chunk := getConfidenceChunk(xEndPos-xStartPos, columns[y][x], columns[y][next])
 			for k, v := range chunk {
 				resultMatrix[y][xStartPos+k] = v
@@ -251,22 +251,22 @@ func ScaleConfidenceMatrix(width int, height int, confidence *[][]float32) [][]f
 	return resultMatrix
 }
 
-func getConfidenceColumns(height int, confidence *[][]float32) [][]float32 {
-	result := make([][]float32, height)
+func getConfidenceColumns(height int, confidence *[][]float64) [][]float64 {
+	result := make([][]float64, height)
 	xSize := len((*confidence)[0])
 	arrEnd := len(*confidence) - 1
 	adjustedHeight := height - 1 // adjusts height to work with arrays
 	// init array
 	for i := range result {
-		result[i] = make([]float32, xSize)
+		result[i] = make([]float64, xSize)
 	}
 	for i := 0; i < arrEnd; i++ {
 		for j := 0; j < xSize; j++ {
 			next := i + 1
 			start := (*confidence)[i][j]                                                      // get start confidence value for chunk
 			end := (*confidence)[next][j]                                                     // get end confidence value for chunk
-			yStartPos := int(lerp(0.0, float32(adjustedHeight), float32(i)/float32(arrEnd)))  // lerp indices to get start index
-			yEndPos := int(lerp(0.0, float32(adjustedHeight), float32(next)/float32(arrEnd))) // lerp indices to get end index
+			yStartPos := int(lerp(0.0, float64(adjustedHeight), float64(i)/float64(arrEnd)))  // lerp indices to get start index
+			yEndPos := int(lerp(0.0, float64(adjustedHeight), float64(next)/float64(arrEnd))) // lerp indices to get end index
 			numElements := yEndPos - yStartPos                                                // calculate number of elements for chunk
 			chunk := getConfidenceChunk(numElements, start, end)
 			for k, v := range chunk {
@@ -277,16 +277,16 @@ func getConfidenceColumns(height int, confidence *[][]float32) [][]float32 {
 	return result
 
 }
-func getConfidenceChunk(numElements int, start float32, end float32) []float32 {
-	result := make([]float32, numElements+1) // adds one to buffer because there is overlap in the scaling of the matrix for example
+func getConfidenceChunk(numElements int, start float64, end float64) []float64 {
+	result := make([]float64, numElements+1) // adds one to buffer because there is overlap in the scaling of the matrix for example
 	// first chunk could be from idx 0-2 (inclusive) the next chunk would start from 2-4 (also inclusive), thus overlap
 	for i := 0; i <= numElements; i++ {
-		result[i] = lerp(start, end, float32(i)/float32(numElements))
+		result[i] = lerp(start, end, float64(i)/float64(numElements))
 	}
 	return result
 }
 // ConfidenceMatrixToImage takes the confidences matrix and a supplied colorScale function and returns an image.
-func ConfidenceMatrixToImage(confidence [][]float32, colorScale func(float64) *color.RGBA, opacity uint8) *image.RGBA {
+func ConfidenceMatrixToImage(confidence [][]float64, colorScale func(float64) *color.RGBA, opacity uint8) *image.RGBA {
 	height := len(confidence)
 	width := len(confidence[0])
 	resultImage := image.NewRGBA(image.Rect(0, 0, width, height))
@@ -522,6 +522,6 @@ func getFilePath(datasetDir string, fileID string, bandLabel string, fileType st
 	return path.Join(datasetDir, fileName)
 }
 
-func lerp(v0 float32, v1 float32, t float32) float32 {
+func lerp(v0 float64, v1 float64, t float64) float64 {
 	return (1.0 - t) * v0 + t * v1;
   }
